Add entry point index.js and use modules for better code organization
