+++ index.html
@@ -44,7 +44,7 @@
     />
     <link rel="manifest" href="icons/favicon/site.webmanifest" />
 
-    <script type="module" src="js/main.js"></script>
+    <script type="module" src="index.js"></script>
     <!-- Script for running Tailwind -->
     <script src="https://cdn.tailwindcss.com"></script>
     <script src="https://cdn.jsdelivr.net/npm/tw-elements/dist/js/index.min.js"></script>
