+++ modules/dynamic-book.js
