--- styles/index.css
@@ -1,183 +1,0 @@
-html {
-  scroll-behavior: smooth;
-  box-sizing: border-box;
-}
-
-*,
-*::before,
-*::after {
-  margin: 0;
-  padding: 0;
-  box-sizing: inherit;
-  transition-duration: 0.25s;
-}
-
-header {
-  display: flex;
-  justify-content: space-between;
-  background-color: transparent;
-}
-
-header img {
-  width: 260px;
-  border-radius: 50%;
-}
-
-/* Navbar section */
-.nav-link a {
-  position: relative;
-  transition:
-    transform 0.25s ease-out,
-    font-weight 0s;
-}
-
-.nav-link a::after {
-  position: absolute;
-  content: '';
-  width: 100%;
-  transform: scaleX(0);
-  height: 0.125rem;
-  bottom: -0.625rem;
-  left: 0;
-  background-color: rgb(239, 246, 255);
-  transition: transform 0.25s ease-out;
-}
-
-.nav-link a:hover::after {
-  transform: scaleX(1);
-  transform-origin: center;
-}
-
-a.topic-active {
-  font-weight: bold;
-}
-
-a.topic-active::after {
-  position: absolute;
-  content: '';
-  width: 100%;
-  transform: scaleX(0);
-  height: 0.125rem;
-  bottom: -0.625rem;
-  left: 0;
-  background-color: rgb(239, 246, 255);
-  animation: underline 0.75s;
-}
-
-@keyframes underline {
-  0% {
-    transform: scaleX(0);
-    transform-origin: center;
-  }
-
-  100% {
-    transform: scaleX(1);
-    transform-origin: center;
-  }
-}
-
-.date-time::before {
-  position: absolute;
-  content: '';
-  width: 90%;
-  transform: scaleX(1);
-  height: 0.125rem;
-  top: -0.625rem;
-  left: 0;
-  background-color: rgb(239, 246, 255);
-}
-
-.date-time::after {
-  position: absolute;
-  content: '';
-  width: 90%;
-  transform: scaleX(1);
-  height: 0.125rem;
-  bottom: -0.625rem;
-  left: 0;
-  background-color: rgb(239, 246, 255);
-}
-
-.dummy-heading {
-  font-size: 0;
-  visibility: hidden;
-}
-
-/* Book container */
-.book-container {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 0.5rem 0;
-  border-bottom: 1px solid lightgray;
-}
-
-.book-container:nth-child(odd) {
-  background-color: lightgray;
-}
-
-.book-container h3 span {
-  font-size: 1rem;
-  font-weight: normal;
-}
-
-.book-container .book-title {
-  font-weight: bold;
-}
-
-.book-container .del-btn svg {
-  transition: stroke 100ms;
-}
-
-.book-container .del-btn:hover svg {
-  stroke: #fff;
-}
-
-.primary-btn {
-  cursor: pointer;
-}
-
-/* Scrollbar section */
-::-webkit-scrollbar {
-  width: 0.5625rem;
-}
-
-::-webkit-scrollbar-track {
-  background: #e5e5e5;
-  border-radius: 0.3125rem;
-}
-
-::-webkit-scrollbar-thumb {
-  background:
-    linear-gradient(
-      45deg,
-      rgba(6, 182, 212, 0.7),
-      rgba(91, 14, 214, 0.7) 100%
-    );
-  border-radius: 0.3125rem;
-}
-
-::-webkit-scrollbar-thumb:hover {
-  background: rgb(120 113 108);
-}
-
-.books-display,
-.books-action,
-.contact-us {
-  animation: fade-in 1s;
-  transition: display 1s ease-out;
-}
-
-@keyframes fade-in {
-  0% {
-    opacity: 0;
-  }
-
-  100% {
-    opacity: 1;
-  }
-}
-
-.hide-section {
-  display: none !important;
-}
